pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "matterkeep"
version = "0.1.0"
description = "Archive your Mattermost message history and media into a self-contained offline HTML archive"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "mattermostdriver>=7.3.2",
    "click>=8.1",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "jinja2>=3.1",
    "mistune>=3.0",
    "keyring>=25.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-mock>=3.14",
    "responses>=0.25",
    "ruff>=0.5",
    "mypy>=1.10",
    "pip-audit>=2.7",
]

[project.scripts]
matterkeep = "matterkeep.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
strict = true
python_version = "3.11"

[tool.pytest.ini_options]
testpaths = ["tests"]
